pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Backend for EstuResidencia"
authors = ["Sergio <esgarciac@unal.edu.co>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.2.4"
djangorestframework = "^3.14.0"
coreapi = "^2.3.3"
dj-database-url = "^2.1.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"
pillow = "^10.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

